fix: CI failures - ratelimit bug, flaky tests, Go version
- Fix ReserveN duration calculation: time.Duration(deficit/rate)
  truncated float to 0. Use time.Duration(deficit/rate * float64(Second))
- Fix flaky TestTCPProxyProxyConnectionFullFlow: close mux to unblock
- Remove invalid TLS tests that panic after dead-code removal
- Set go.mod to go 1.23 to match CI runner

+++ internal/ratelimit/ratelimit.go
@@ -99,7 +99,7 @@ func (l *Limiter) ReserveN(n int) time.Duration {
 	}
 
 	deficit := float64(n) - l.tokens
-	waitDuration := time.Duration(deficit/l.rate) * time.Second
+	waitDuration := time.Duration(deficit / l.rate * float64(time.Second))
 
 	return waitDuration
 }
